test(ai): shared openRawAgent + fixture-derived bridge port for WriteGuard e2es (#13330)
Batches the non-blocking test-quality follow-ups from the #13294 + #13325 review threads into one cleanup (no runtime behavior change):

- extract the byte-identical openRawAgent into test/playwright/util/rawAgent.mjs; both WriteGuard NL e2es import the one shared helper (DRY, rule-of-three).
- derive the bridge port from the neuralLink fixture (aiConfig.port) instead of a hardcoded :8081, so raw-ws writers target the same Bridge the fixture uses (un-gated now that the fresh-bridge isolation merged).
- finally-guard BOTH raw-ws writers in the disconnect-release spec (writer-1 leaked if the deny assertion threw before its inline close()).
- carry the deny reason in the admit-poll's value so a timeout names WHY it stayed denied in the expected/received diff.
- reword the multi-writer @summary from the dated LIVE-VERIFIED snapshot to present-tense.
